update components from vue1 to vue2
