--- src/components/account/account.html
@@ -1,39 +1,0 @@
-<transition name="modal">
-<div class="modal-mask" @click="close">
-        <div style="height:30%"></div>
-        <div class="modal-container" @click.stop>
-            <span @click="close" tabindex="0" v-on:keyup.enter="close" aria-label="close" class="close">&times;</span>
-
-          <div class="modal-header">
-            <h3>Delete Account</h3>
-          </div>
-
-          <div class="modal-body">
-              <h4>If you choose to proceed you will lose access to your account and data!</h4>
-              <h4>This action is not reversible</h4>
-              <h4>You must enter your password to confirm you want to delete your account and all your data</h4>
-            <div v-bind:class="errorClass">
-                <label v-if="isError">{{ error }}</label>
-                <div class="form-group">
-                    <input :type="passwordFieldType" name="password1" id="password1" class="form-control password" v-model="password1" placeholder="Existing password">
-                    <span v-bind:class="['fa', 'password' == passwordFieldType ? 'fa-eye password-eye' : 'fa-eye-slash password-eye-slash']" @click="togglePassword1()"></span>
-                </div>
-            </div>
-          </div>
-        <warning
-                v-if="showWarning"
-                v-on:hide-warning="showWarning = false"
-                :warning_message='warning_message'
-                :warning_body="warning_body"
-                :consumer_func="warning_consumer_func">
-        </warning>
-      <div class="modal-footer">
-          <slot name="footer">
-          <button class="btn btn-danger" @click="deleteAccountAction">
-              Delete Account
-          </button>
-          </slot>
-      </div>
-    </div>
-</div>
-</transition>
