+++ src/components/choice/Choice.vue
@@ -1,3 +1,4 @@
+<template>
 <transition name="modal">
 <div class="modal-mask" @click="close">
   <div style="height:30%"></div>
@@ -27,3 +28,27 @@
   </div>
 </div>
 </transition>
+</template>
+<script>
+module.exports = {
+    data: function() {
+        return {
+            picked: 0
+        }
+    },
+    props: ['choice_message', 'choice_body', 'choice_options', 'choice_consumer_func'],
+    created: function() {
+    },
+    methods: {
+        close: function() {
+            this.$emit("hide-choice");
+        },
+        confirm: function() {
+            this.close();
+            this.choice_consumer_func(this.picked);
+        }
+    }
+}
+</script>
+<style>
+</style>
